refactor: move node-zendesk types to @connectors/@types (consistent with talisman)
