+++ connectors/src/@types/node-zendesk.d.ts
